+++ docs/reference/changelog.rst
@@ -18,6 +18,12 @@ Current version
 
 
 
+v5.5.1 - 2022-07-26
+-------------------
+
+- ``Fixed`` Unable to configure dropbox backup - Dropbox SDK downgrade [`#1677 <https://github.com/dsmrreader/dsmr-reader/issues/1677>`_]
+
+
 v5.5.0 - 2022-07-15
 -------------------